Backend/requirement.txt
fastapi
uvicorn
openai
chromadb
python-dotenv
numpy
youtube-transcript-api
PyMuPDF
